+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Update `icupy.icu.ResourceBundle.get_binary()` to return `memoryview` instead of `bytes` ([#213])
 - Update `icupy.icu.UnicodeString.get_buffer()` and `.get_terminated_buffer()` to return `memoryview` instead of `ConstChar16Ptr` ([#215])
 - Update docstring ([#231])
+- Add rich comparison operators for `icupy.icu.CollationKey` ([#235])
 - Deprecate `icupy.icu.Transliterator.register_instance()` and `.unregister()` ([#221])
 - Deprecate `icupy.icu.ConstVoidPtr`; use `icupy.icu.UserContext` instead ([#227])
 - Deprecate `icupy.icu.BreakIterator.BreakIterator.DONE` and `icupy.icu.BreakIterator.DONE`; use `UBRK_DONE` instead ([#229])
@@ -383,3 +384,4 @@ Initial release.
 [#231]: https://github.com/miute/icupy/pull/231
 [#232]: https://github.com/miute/icupy/pull/232
 [#234]: https://github.com/miute/icupy/pull/234
+[#235]: https://github.com/miute/icupy/pull/235